Move entity-info.decorator.ts to dependencies dir
