+++ packages/api/cms-api/src/dam/files/entities/file.entity.ts
@@ -16,8 +16,8 @@ import { Type } from "@nestjs/common";
 import { Field, ID, Int, ObjectType } from "@nestjs/graphql";
 import { v4 as uuid } from "uuid";
 
+import { EntityInfo } from "../../../dependencies/decorators/entity-info.decorator";
 import { DamScopeInterface } from "../../types";
-import { EntityInfo } from "../decorators/entity-info.decorator";
 import { FilesEntityInfoService } from "../files-entity-info.service";
 import { DamFileImage } from "./file-image.entity";
 import { FolderInterface } from "./folder.entity";
